:trollface: package responsibility
