+++ stream-plugin/stream-plugin-mybatis-plus/src/main/java/io/github/vampireachao/stream/plugin/mybatisplus/injector/IMapper.java
@@ -1,13 +1,15 @@
 package io.github.vampireachao.stream.plugin.mybatisplus.injector;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import io.github.vampireachao.stream.plugin.mybatisplus.injector.constant.PluginConst;
 import org.apache.ibatis.annotations.Param;
 
 import java.util.Collection;
 
-import static io.github.vampireachao.stream.plugin.mybatisplus.injector.SqlInjectorConfig.COLLECTION_PARAM_NAME;
 
 /**
+ * 拓展BaseMapper
+ *
  * @author VampireAchao
  */
 public interface IMapper<T> extends BaseMapper<T> {
@@ -18,6 +20,6 @@ public interface IMapper<T> extends BaseMapper<T> {
      * @param list 数据
      * @return 条数
      */
-    long insertOneSql(@Param(COLLECTION_PARAM_NAME) Collection<T> list);
+    long insertOneSql(@Param(PluginConst.COLLECTION_PARAM_NAME) Collection<T> list);
 
 }
