+++ stream-plugin/stream-plugin-mybatis-plus/src/main/java/io/github/vampireachao/stream/plugin/mybatisplus/injector/StreamQuerySqlInjectorAutoConfiguration.java
@@ -1,0 +1,33 @@
+package io.github.vampireachao.stream.plugin.mybatisplus.injector;
+
+import com.baomidou.mybatisplus.core.injector.AbstractMethod;
+import com.baomidou.mybatisplus.core.injector.DefaultSqlInjector;
+import com.baomidou.mybatisplus.core.metadata.TableInfo;
+import io.github.vampireachao.stream.plugin.mybatisplus.injector.methods.InsertOneSql;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+
+import java.util.List;
+
+/**
+ * MPSql注入
+ *
+ * @author VampireAchao
+ */
+public class StreamQuerySqlInjectorAutoConfiguration {
+
+
+    @Bean
+    @ConditionalOnMissingBean(DefaultSqlInjector.class)
+    public DefaultSqlInjector defaultSqlInjector() {
+        return new DefaultSqlInjector() {
+            @Override
+            public List<AbstractMethod> getMethodList(Class<?> mapperClass, TableInfo tableInfo) {
+                List<AbstractMethod> methodList = super.getMethodList(mapperClass, tableInfo);
+                methodList.add(new InsertOneSql(SqlMethodEnum.INSERT_ONE_SQL.getMethod()));
+                return methodList;
+            }
+        };
+    }
+
+}
