+++ stream-plugin/stream-plugin-mybatis-plus/src/main/java/io/github/vampireachao/stream/plugin/mybatisplus/injector/constant/PluginConst.java
@@ -1,0 +1,12 @@
+package io.github.vampireachao.stream.plugin.mybatisplus.injector.constant;
+
+/**
+ * @author VampireAchao
+ */
+public interface PluginConst {
+
+    String DEFAULT = "default";
+    String NON_NULL_CONDITION = "%s != null and %s != null";
+    String COLLECTION_PARAM_NAME = "list";
+
+}
