+++ stream-plugin/stream-plugin-mybatis-plus/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.github.vampireachao.stream.plugin.mybatisplus.injector.SqlInjectorConfig
+io.github.vampireachao.stream.plugin.mybatisplus.injector.StreamQuerySqlInjectorAutoConfiguration
