+++ stream-plugin/stream-plugin-mybatis-plus/src/test/java/io/github/vampireachao/stream/plugin/mybatisplus/mapper/UserInfoMapper.java
@@ -1,12 +1,12 @@
 package io.github.vampireachao.stream.plugin.mybatisplus.mapper;
 
 import io.github.vampireachao.stream.plugin.mybatisplus.injector.IMapper;
+import io.github.vampireachao.stream.plugin.mybatisplus.injector.methods.InsertOneSql;
 import io.github.vampireachao.stream.plugin.mybatisplus.pojo.po.UserInfo;
 import org.apache.ibatis.annotations.Param;
 
 import java.util.Collection;
 
-import static io.github.vampireachao.stream.plugin.mybatisplus.injector.SqlInjectorConfig.COLLECTION_PARAM_NAME;
 
 /**
  * UserInfoMapper
@@ -16,5 +16,5 @@ import static io.github.vampireachao.stream.plugin.mybatisplus.injector.SqlInjec
  */
 public interface UserInfoMapper extends IMapper<UserInfo> {
 
-    long updateOneSql(@Param(COLLECTION_PARAM_NAME) Collection<UserInfo> list);
+    long updateOneSql(@Param(InsertOneSql.COLLECTION_PARAM_NAME) Collection<UserInfo> list);
 }
